+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.26.2] - 2025-02-27
+### Added
+- Added `api.customScript` for management of Custom Scripts
+
 ## [1.26.1] - 2025-02-24
 ### Fixed
 - Dependency on non existing package `@signageos/forward-server-bridge@0.0.1`